requirements.txt
flask
flask-cors
requests
language_tool_python
gunicorn

# 🔊 Transcripción con Whisper
openai-whisper
torch
numpy